renaming file and component names
