+++ apps/meteor/client/views/room/modals/ForwardMessageModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ForwardMessageModal';
