--- apps/meteor/client/views/room/modals/ShareMessageModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ShareMessageModal';
